Changed Decimals to encode to JSON numbers not strings
Strings are still accepted on decode. Note that Go cannot preserve precision with this representation, as the JSON parser funnels JSON numbers through the float64 type. But that is a Go bug. Java BigDecimals and Jackson handle it fine. So I opted to use a number, which seems more correct.

Note that using Decimal literals in SADL work fine, i.e. default values for a Decimal field, or for example literals. So the spec is fine, and codegen for Java should be fine, but Go implementation will lose precision until the JSON parser if fixed (if ever).